ActiveDecorator導入 #39
