+++ app/models/profile.rb
@@ -20,15 +20,4 @@ class Profile < ApplicationRecord
   belongs_to :user
   has_one_attached :avatar
 
-  #年齢計算
-  def age_cal
-    return '不明' unless age.present?
-    years = Time.zone.now.year - age.year
-    days = Time.zone.now.yday - age.yday
-    if days < 0
-      "#{years - 1}歳"
-    else
-      "#{years}歳"
-    end
-  end
 end
